Add files via upload
All the backgrounds! Might have to tweak one or two based on what happens with the GUI, but those won't be major changes. So this is more or less the official final BGs (unless I add lineart to the battle arena cause now it's slightly out of place compared to others...).